Week 7 Commit and contribution
This week we have mainly determined what aspects of the game's player decisions should include. In the game, the player's key decisions mainly revolve around seven aspects, which directly affect the direction and ending of the story. This is just an outline, and we will write in detail about the specific decision content of each aspect later.