Week 7 commit and contribution
This could be the first version of "Rulebook", and I wrote this based on the modified version of “Game Creation: Design Document”. I wrote a brief background of the game, the game mechanics and the whole process. It's very simple because I don't think we need too much narrative, it gets complicated to the players, but you guys can modify it if you have new ideas. Rita week 7