Week 8 commit and contribution
This is the first version of story recap. I wrote this using the draft that PJ done. This doc includes the basic information including Background, The Three Nations, The Chosen Council, The Sacred Lion Succession Ceremony, Key Player Characters, Themes and Storyline. But this is not a final one, we should keep improving.