Week 9 Commit and contribution
This week I created an outline for the different endings, and we discussed and designed four endings. I summarized the endings in simple words, and PJ will be responsible for the subsequent refinement. In addition, I further described the character development elements of each character in the character booklet. The specific content of the characters was also refined by PJ. In the next week, Rita and I will review and modify the content provided by PJ, and start to make the actual game booklet. We will reorganize and revise all the documents in week 10.