Week 10 Commit and Contribution
I went over all the documents again this week. I changed the font of all the documents to make them easier for the player to read. I have also added some notes to the document according to the order in which the book is used. I merged the ending chart and ending documents. Next, Rita and I will add a cover to the character book.