Week 11 Commit & Contribution
This week I wrote the game flow and the voiceover script for the game video, and printed and organized all the content. I invited five friends to play our game, I introduced the game to them, recorded their game process and edited it. Finally, I uploaded the video to Google Drive and uploaded the link to GitHub as an md file.