Add files via upload
I made a presentation PowerPoint based on the video we shot this week for the demo presentation on Thursday. The text contains the background of the game, the characters and the exact gameplay flow. And I used screenshots from the video to show each step. I aslo list our materials of the game. We should take the actual one on lecture.